+++ NEWS.md
@@ -6,7 +6,7 @@
 
 ## Breaking Changes
 - major rework: no subplotting anymore -> now everything is in a single plot and can be zoomed seamlessly when using the plotly option
-- events are drawn from top to bottom (not from bottom to top) as given in input data frame. This renders the result more intuitive (when using `optimize_y = FALSE`).
+- events are drawn from top to bottom (not from bottom to top) as given in input data frame. This renders the result more intuitive.
 
 ## Minor changes
 - arguments `showLabels` and `lineInterval` have long been deprecated and have now been removed (use `show_labels` and `background_lines`)